Reorganize html-table tests by table class
Split tests/test_html_tables_to_json.py into four files aligned with
the A/B/C taxonomy from issue #13:

- test_class_a_tables.py (headers in first row)
- test_class_b_tables.py (headers in first column)
- test_class_c_tables.py (headless)
- test_html_tables_misc.py (multi-table / thead, span classes)

Adds edge-case coverage for empty cells, variable column counts, and
single-column class A tables. Coverage on convert_html_tables.py rises
from 97% to 100%.
